refactor: CustomInput -> Input 파일명 및 스타일 수정
